解决loading闪屏处理

+++ src/containers/HZLB/Controller.js
@@ -1,4 +1,4 @@
-import {mapActions} from 'vuex';
+import {mapActions, mapMutations} from 'vuex';
 
 const controller = {
   init(mescroll) {
@@ -35,29 +35,32 @@ const controller = {
     }
   },
   exeDeleteVisitorList(params) {
-    params = {value: params};
-    this.deleteVisitorList(params)
-      .then((res) => {
-        res = res || {};
-        const {success} = res;
-        if (success) {
-          this.$vux.toast.show({
-            text: '操作成功'
-          });
-        } else {
+    const result = confirm('确认解绑就诊卡吗?');
+    if (result) {
+      params = {value: params};
+      this.deleteVisitorList(params)
+        .then((res) => {
+          res = res || {};
+          const {success} = res;
+          if (success) {
+            this.$vux.toast.show({
+              text: '操作成功'
+            });
+          } else {
+            this.$vux.toast.show({
+              type: 'cancel',
+              text: '操作失败'
+            });
+          }
+        })
+        .catch((err) => {
+          console.log(err);
           this.$vux.toast.show({
             type: 'cancel',
             text: '操作失败'
           });
-        }
-      })
-      .catch((err) => {
-        console.log(err);
-        this.$vux.toast.show({
-          type: 'cancel',
-          text: '操作失败'
         });
-      });
+    }
   },
   exeUpdateVisitorList(params) {
     params = {value: params};
@@ -105,15 +108,14 @@ const controller = {
         this.exeCommonCallback();
       });
   },
-  exeClearVisitorList() {
-    this.clearVisitorList();
-  },
   ...mapActions([
     'insertVisitorList',
     'deleteVisitorList',
     'updateVisitorList',
-    'selectVisitorList',
-    'clearVisitorList'
+    'selectVisitorList'
+  ]),
+  ...mapMutations([
+    'CLEAR_VISITOR_LIST_SUCCESS'
   ])
 };
